Implemented ranged attack, bullet damage scales with level of dexterity When in ranged state animations are now dependent on mousePos ToDo - fix mousePos being calculated from 0,0,0 instead of playerPos

+++ Assets/+++Workdata/Scripts/Player/PlayerInfos.cs
@@ -108,7 +108,7 @@ public class PlayerInfos : MonoBehaviour, IDataPersistence
         else
             inventoryState = 0;
 
-        playerController.anim.SetInteger("InventoryState", inventoryState);
+        playerController.anim.SetFloat("InventoryState", inventoryState);
     }
 
     //calculates player health with the damage values from the EnemyInfos.cs script
